+++ test/vissen/output/pixel_buffer_test.rb
@@ -5,11 +5,28 @@ require 'test_helper'
 describe Vissen::Output::PixelBuffer do
   subject { Vissen::Output::PixelBuffer }
 
-  let(:rows)         { 6 }
-  let(:columns)      { 8 }
-  let(:grid_context) { Vissen::Output::Context::Grid.new rows, columns }
+  let(:rows)          { 6 }
+  let(:columns)       { 8 }
+  let(:context_klass) { Vissen::Output::Context::Grid }
+  let(:grid_context)  { context_klass.new rows, columns }
+  let(:filter_klass)  { Vissen::Output::Filter::Gamma }
+  let(:filter)        { filter_klass.new grid_context }
+  let(:pixel_buffer)  { subject.new grid_context }
 
-  let(:pixel_buffer) { subject.new grid_context }
+  describe '.new' do
+    it 'accepts an optional array of fitlers' do
+      filters = [filter]
+      assert_silent { subject.new grid_context, filters }
+    end
+    
+    it 'raises a TypeError if any of the filters does not share context' do
+      other_context = context_klass.new 1, 2
+      other_filter  = filter_klass.new other_context
+      filters = [filter, other_filter]
+      
+      assert_raises(TypeError) { subject.new grid_context, filters }
+    end
+  end
 
   describe '#copy!' do
     it 'copies the pixels from another pixel_buffer' do
@@ -43,6 +60,34 @@ describe Vissen::Output::PixelBuffer do
         assert_equal 0, pixel.r
         assert_equal 0, pixel.g
         assert_equal 0, pixel.b
+      end
+    end
+  end
+
+  describe '#finalize!' do
+    it 'applies the output filters' do
+      pixel_buffer = subject.new grid_context, [filter]
+      
+      r = rand
+      g = rand
+      b = rand
+      
+      pixel_buffer.pixels.each do |pixel|
+        pixel.r = r
+        pixel.g = g
+        pixel.b = b
+      end
+      
+      r_gamma = r**filter.value
+      g_gamma = g**filter.value
+      b_gamma = b**filter.value
+      
+      pixel_buffer.finalize!
+      
+      pixel_buffer.pixels.each do |pixel|
+        assert_in_epsilon r_gamma, pixel.r
+        assert_in_epsilon g_gamma, pixel.g
+        assert_in_epsilon b_gamma, pixel.b
       end
     end
   end